+++ app/assets/scss/main.scss
@@ -57,3 +57,7 @@
     }
   }
 }
+
+.no-scrollbar::-webkit-scrollbar{
+  display: none;
+}
