+++ app/pages/marketplace.vue
@@ -1,7 +1,6 @@
 <template>
     <Title>MarketPlace</Title>
     <div class="flex min-h-screen w-full -mt-3">
-        <Aside />
         <div class="flex-1 flex justify-center">
             <h1 class="font-medium text-lg mt-10">Coming Soon</h1>
         </div>
