+++ app/pages/pricing.vue
@@ -1,3 +1,5 @@
 <template>
-    
+    <div class="w-full h-full flex-1 flex justify-center">
+        <h1 class="font-medium text-lg mt-10">Coming Soon</h1>
+    </div>
 </template>
