+++ server/db/drizzle/0014_reflective_masked_marvel.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "forms" ADD COLUMN "price_group_message" text;
